# docs/openapi.yaml
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/admin/login": {
      "post": {
        "description": "Аутентификация администратора по email и паролю. Возвращает JWT токен и информацию о пользователе. Токен сохраняется в БД (механизм one-device-one-session).",
        "operationId": "AuthController_loginAdmin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginAdminDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Успешная авторизация",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "access_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJjYjc5OTRmZS02MmNlLTRhOGQtOWQ5NS0zYjA5NWE1MTBkNDEiLCJlbWFpbCI6InJvb3RAZXhhbXBsZS5jb20iLCJyb2xlIjoiUk9PVCIsImlhdCI6MTYwMDAwMDAwMCwiZXhwIjoxNjAwNjA0ODAwfQ.signature",
                    "admin": {
                      "id": "cb7994fe-62ce-4a8d-9d95-3b095a510d41",
                      "name": "Root Admin",
                      "email": "root@example.com",
                      "role": "ROOT"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Неверные учетные данные",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Invalid email or password",
                    "error": "Unauthorized",
                    "statusCode": 401
                  }
                }
              }
            }
          }
        },
        "summary": "Вход администратора",
        "tags": [
          "Аутентификация"
        ]
      }
    },
    "/auth/shop/login": {
      "post": {
        "description": "Аутентификация магазина по логину и паролю. Возвращает JWT токен и информацию о магазине. Токен сохраняется в БД.",
        "operationId": "AuthController_loginShop",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginShopDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Успешная авторизация",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "access_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJhOTZlN2JhZC1kNGQ4LTQ1NTktYjliNy0wZTE4MGZlZGQzZTMiLCJsb2dpbiI6InRlc3RzaG9wMSIsImlhdCI6MTYwMDAwMDAwMCwiZXhwIjoxNjAwNjA0ODAwfQ.signature",
                    "shop": {
                      "id": "a96e7bad-d4d8-4559-b9b7-0e180fedd3e3",
                      "name": "Test Shop",
                      "login": "testshop1",
                      "address": "123 Test St",
                      "inn": "1234567890"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Неверные учетные данные",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Invalid login or password",
                    "error": "Unauthorized",
                    "statusCode": 401
                  }
                }
              }
            }
          }
        },
        "summary": "Вход магазина",
        "tags": [
          "Аутентификация"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "description": "Завершение текущей сессии. Инвалидирует JWT токен (удаляет из БД). Требует авторизацию.",
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Успешный выход из системы",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Вы успешно вышли из системы"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Unauthorized",
                    "statusCode": 401
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Выход из системы",
        "tags": [
          "Аутентификация"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "description": "Обновление истекающего или истекшего JWT токена. Возвращает новый токен с обновленным временем жизни.",
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Токен успешно обновлен",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "access_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJjYjc5OTRmZS02MmNlLTRhOGQtOWQ5NS0zYjA5NWE1MTBkNDEiLCJlbWFpbCI6InJvb3RAZXhhbXBsZS5jb20iLCJyb2xlIjoiUk9PVCIsImlhdCI6MTYwMDAwMDAwMCwiZXhwIjoxNjAwNjA0ODAwfQ.new_signature",
                    "user": {
                      "id": "cb7994fe-62ce-4a8d-9d95-3b095a510d41",
                      "email": "root@example.com",
                      "role": "ROOT"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Недействительный токен",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Invalid token",
                    "error": "Unauthorized",
                    "statusCode": 401
                  }
                }
              }
            }
          }
        },
        "summary": "Обновление JWT токена",
        "tags": [
          "Аутентификация"
        ]
      }
    },
    "/admins": {
      "get": {
        "operationId": "AdminsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Список администраторов"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Получить список администраторов",
        "tags": [
          "Администраторы"
        ]
      },
      "post": {
        "description": "Создание нового администратора только с ролью MANAGER. Доступно только ROOT.",
        "operationId": "AdminsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAdminDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Администратор успешно создан",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "cb7994fe-62ce-4a8d-9d95-3b095a510d41",
                    "name": "Имя Фамилия",
                    "email": "manager@example.com",
                    "role": "MANAGER",
                    "createdAt": "2025-01-17T10:00:00.000Z",
                    "updatedAt": "2025-01-17T10:00:00.000Z"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Некорректные данные или email уже используется",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Email already exists",
                    "error": "Bad Request",
                    "statusCode": 400
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Создать нового администратора",
        "tags": [
          "Администраторы"
        ]
      }
    },
    "/admins/{id}/password": {
      "patch": {
        "operationId": "AdminsController_updatePassword",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Пароль изменен"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Изменить пароль администратора (только ROOT)",
        "tags": [
          "Администраторы"
        ]
      }
    },
    "/admins/{id}": {
      "delete": {
        "operationId": "AdminsController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Администратор удален"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Удалить администратора (только ROOT)",
        "tags": [
          "Администраторы"
        ]
      }
    },
...
    "/profile/password": {
      "patch": {
        "description": "Изменение пароля для текущего авторизованного пользователя (администратора или магазина). Требует подтверждение текущего пароля. Для администраторов: токен автоматически инвалидируется (удаляется из БД) для безопасности. Для магазинов: пароль обновляется в БД. После смены пароля требуется повторный вход.",
        "operationId": "ProfileController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Пароль успешно изменен",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Пароль успешно изменен. Пожалуйста, войдите заново."
                  }
                }
              }
            }
          },
          "401": {
            "description": "Неверный текущий пароль или пользователь не авторизован",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Неверный текущий пароль",
                    "error": "Unauthorized",
                    "statusCode": 401
                  }
                }
              }
            }
          },
          "404": {
            "description": "Пользователь не найден",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Администратор не найден",
                    "error": "Not Found",
                    "statusCode": 404
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Смена пароля текущего пользователя",
        "tags": [
          "Профиль"
        ]
      }
    }
  },
  "info": {
    "title": "Shop Terminals Management API",
    "description": "",
    "version": "1.0",
    "contact": {

    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Введите JWT токен полученный при авторизации",
        "in": "header"
      }
    },
    "schemas": {
      "LoginAdminDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "root@example.com",
            "description": "Email администратора"
          },
          "password": {
            "type": "string",
            "example": "rootpass123",
            "description": "Пароль (минимум 6 символов)"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LoginShopDto": {
        "type": "object",
        "properties": {
          "login": {
            "type": "string",
            "example": "shop1",
            "description": "Логин магазина"
          },
          "password": {
            "type": "string",
            "example": "shoppass123",
            "description": "Пароль (минимум 6 символов)"
          }
        },
        "required": [
          "login",
          "password"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
            "description": "JWT токен для обновления"
          }
        },
        "required": [
          "token"
        ]
      },
      "CreateAdminDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Имя Фамилия",
            "description": "Полное имя администратора"
          },
          "email": {
            "type": "string",
            "example": "manager@example.com",
            "description": "Email администратора, unique"
          },
          "password": {
            "type": "string",
            "example": "123456",
            "description": "Пароль (минимум 6 символов)"
          }
        },
        "required": [
          "name",
          "email",
          "password"
        ]
      },
      "UpdatePasswordDto": {
        "type": "object",
        "properties": {
          "newPassword": {
            "type": "string",
            "example": "newsecurepass456",
            "description": "Новый пароль (минимум 6 символов)"
          }
        },
        "required": [
          "newPassword"
        ]
      },
...
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string",
            "example": "changeMe",
            "description": "Текущий пароль"
          },
          "newPassword": {
            "type": "string",
            "example": "newPassword",
            "description": "Новый пароль (минимум 6 символов)"
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ]
      }
    }
  }
}